feat: Extract inline tests from prediction.rs into tests/prediction_unit_tests.rs
